I've made some changes to how I search for satellites.
Based on your feedback, I've adjusted my approach to satellite searching. Now, I can:
1. Search satellites based on a query string.
2. Search satellites based on a category ID.

This change aims to make searching more straightforward and focused for each type of query.
